refactor: reorganize folder structure and update imports across `user-service`
- Moved `Command`, `Validators`, and `DTO` files to appropriate `Domain`/`Application` subdirectories.
- Adjusted imports to align with the new structure.
- Removed debug `console.log` statements from `CreateUserCommandHandler`.
