+++ game-service/src/Infrastructure/Http/Concrete/BackendApiClient.ts
@@ -31,7 +31,7 @@ export class BackendApiClient implements IBackendApiClient
         }
     }
 
-    //TODO: implementar
+    //Não implementado
     public async VerifyIfUserExists(uuid: string): Promise<boolean>
     {
         try
@@ -51,6 +51,7 @@ export class BackendApiClient implements IBackendApiClient
         }
     }
 
+    //TODO: implementar
     public async VerifyIfUserExistsByUsername(username: string): Promise<boolean>
     {
         try
