+++ user-service/src/Domain/Command/Validators/BaseValidator.ts
