+++ user-service/src/Domain/Command/Validators/CreateUserCommandValidator.ts
@@ -1,7 +1,7 @@
 import {CreateUserCommand} from "../CommandObject/CreateUserCommand.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
-import {EmailVO} from "../../../Domain/ValueObjects/EmailVO.js";
-import {PasswordHashVO} from "../../../Domain/ValueObjects/PasswordHashVO.js";
+import {EmailVO} from "../../ValueObjects/EmailVO.js";
+import {PasswordHashVO} from "../../ValueObjects/PasswordHashVO.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
 import {BaseValidator} from "./BaseValidator.js";
 import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
