+++ user-service/src/Domain/Queries/Handlers/BaseHandlerQuery.ts
